requirements.txt
nicegui>=2.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
pandas>=2.0
plotly>=5.0
openpyxl>=3.1
python-dotenv>=1.0